+++ deployment_example/create_exe.py
@@ -4,10 +4,11 @@
 Deployment example
 ==================
 
-Deployment script using `guidata.disthelpers` (py2exe or cx_Freeze)
+Deployment script using `guidata.utils.disthelpers` (py2exe or cx_Freeze)
 """
 
-from guidata import disthelpers as dh
+from guidata.utils import disthelpers as dh
+
 
 def create_exe():
     dist = dh.Distribution()
@@ -15,6 +16,6 @@ def create_exe():
     dist.add_modules('guidata', 'guiqwt')
     dist.build_cx_freeze()  # use `build_py2exe` to use py2exe instead
 
+
 if __name__ == '__main__':
     create_exe()
-