+++ guiqwt/curve.py
@@ -115,7 +115,8 @@ from qtpy.QtWidgets import (
 from qtpy.QtGui import QBrush, QColor, QPen, QPolygonF
 from qtpy.QtCore import Qt, QPointF, QLineF, QRectF, Signal
 
-from guidata.utils import assert_interfaces_valid, update_dataset
+from guidata.utils.misc import assert_interfaces_valid
+from guidata.utils import update_dataset
 from guidata.configtools import get_icon, get_image_layout
 from guidata.qthelpers import create_action, add_actions