+++ guiqwt/histogram.py
@@ -51,7 +51,8 @@ from qtpy.QtWidgets import QHBoxLayout, QVBoxLayout, QToolBar
 
 from guidata.dataset.datatypes import DataSet
 from guidata.dataset.dataitems import FloatItem
-from guidata.utils import assert_interfaces_valid, update_dataset
+from guidata.utils.misc import assert_interfaces_valid
+from guidata.utils import update_dataset
 from guidata.configtools import get_icon, get_image_layout
 from guidata.qthelpers import add_actions, create_action