+++ guiqwt/shapes.py
@@ -79,7 +79,8 @@ import sys
 from math import cos, fabs, pi, sin, sqrt
 
 import numpy as np
-from guidata.utils import assert_interfaces_valid, update_dataset
+from guidata.utils.misc import assert_interfaces_valid
+from guidata.utils import update_dataset
 from qtpy.QtCore import QLineF, QPointF, QRectF, Qt
 from qtpy.QtGui import QBrush, QPainter, QPen, QPolygonF, QTransform